# Conflicts:
#	include/fvad.h
#	src/fvad.c
#	src/signal_processing/signal_processing_library.h
#	src/signal_processing/spl_inl.h
#	tests/signal_processing_unittest.c
#	tests/vad_sp_unittest.c
#	tests/vad_unittest.c
#	tests/vad_unittest.h

+++ tests/vad_filterbank_unittest.c
@@ -29,8 +29,8 @@ void test_main() {
   // Construct a speech signal that will trigger the VAD in all modes. It is
   // known that (i * i) will wrap around, but that doesn't matter in this case.
   int16_t speech[kMaxFrameLength];
-  for (int16_t i = 0; i < kMaxFrameLength; ++i) {
-    speech[i] = (i * i);
+  for (size_t i = 0; i < kMaxFrameLength; ++i) {
+    speech[i] = i * i;
   }
 
   int frame_length_index = 0;
@@ -64,7 +64,7 @@ void test_main() {
 
   // Verify that all ones in gives kOffsetVector out. Any other constant input
   // will have a small impact in the sub bands.
-  for (int16_t i = 0; i < kMaxFrameLength; ++i) {
+  for (size_t i = 0; i < kMaxFrameLength; ++i) {
     speech[i] = 1;
   }
   for (size_t j = 0; j < kFrameLengthsSize; ++j) {
