+++ tests/vad_sp_unittest.c
@@ -15,6 +15,7 @@
 void test_main() {
   VadInstT* self = malloc(sizeof(VadInstT));
   #define kMaxFrameLenSp 960  // Maximum frame length in this unittest.
+
   int16_t zeros[kMaxFrameLenSp] = { 0 };
   int32_t state[2] = { 0 };
   int16_t data_in[kMaxFrameLenSp];
@@ -31,14 +32,14 @@ void test_main() {
 
   // Construct a speech signal that will trigger the VAD in all modes. It is
   // known that (i * i) will wrap around, but that doesn't matter in this case.
-  for (int16_t i = 0; i < kMaxFrameLenSp; ++i) {
-    data_in[i] = (i * i);
+  for (size_t i = 0; i < kMaxFrameLenSp; ++i) {
+    data_in[i] = i * i;
   }
   // Input values all zeros, expect all zeros out.
   WebRtcVad_Downsampling(zeros, data_out, state, kMaxFrameLenSp);
   EXPECT_EQ(0, state[0]);
   EXPECT_EQ(0, state[1]);
-  for (int16_t i = 0; i < kMaxFrameLenSp / 2; ++i) {
+  for (size_t i = 0; i < kMaxFrameLenSp / 2; ++i) {
     EXPECT_EQ(0, data_out[i]);
   }
   // Make a simple non-zero data test.
