+++ tests/vad_unittest.h
@@ -23,12 +23,13 @@ static const size_t kRatesSize = sizeof(kRates) / sizeof(*kRates);
 
 // Frame lengths we support.
 #define kMaxFrameLength 1440
-static const int kFrameLengths[] = { 80, 120, 160, 240, 320, 480, 640, 960,
+static const size_t kFrameLengths[] = { 80, 120, 160, 240, 320, 480, 640, 960,
+
     kMaxFrameLength };
 static const size_t kFrameLengthsSize = sizeof(kFrameLengths) / sizeof(*kFrameLengths);
 
 
   // Returns true if the rate and frame length combination is valid.
-  bool ValidRatesAndFrameLengths(int rate, int frame_length);
+  bool ValidRatesAndFrameLengths(int rate, size_t frame_length);
 
 #endif  // WEBRTC_COMMON_AUDIO_VAD_VAD_UNITTEST_H
